Synchro and j3 installer
git-svn-id: https://dev.virtuemart.net/svn/virtuemart/trunk/virtuemart@9201 87cc427b-4a1e-0410-b729-b2260ec0f063

+++ administrator/components/com_virtuemart/helpers/vmviewadmin.php
@@ -89,10 +89,10 @@ class VmViewAdmin extends vView {
 		if(vmAccess::manager($view.'.edit')){
 			vToolBarHelper::editList();
 		}
-		if(vmAccess::manager($view.'.create')){
+		if($showNew and vmAccess::manager($view.'.create')){
 			vToolBarHelper::addNew();
 		}
-		if(vmAccess::manager($view.'.delete')){
+		if($showDelete and vmAccess::manager($view.'.delete')){
 			vToolBarHelper::spacer('10');
 			vToolBarHelper::deleteList();
 		}
